removed fsanitize, conditioned MAKE scripts so they only come up when necessary, protected conf file

+++ webserv_sweatZer/scripts/init.sh
@@ -11,48 +11,54 @@ options=("y" "n")
 
 echo -e "\n${CYAN}    -----> Initialising the infamous SweatZer Webserv... <-----\n${NC}"
 
-while true; do
-    read -p $'\033[01;35mDo you wish to install php-cgi (only available on Ubuntu w/ sudo rights)?: y/n \033[0m' yn
-    case $yn in
-        [Yy]* ) if [[ "$OSTYPE" == "linux-gnu"* ]] 
-                then
-                    echo -e "Installing php-cgi..." &&
-                    sudo apt-get install php-cgi &&
-                    echo -e "php-cgi installed!"
-                else
-                    echo -e "${RED}Please run this webserv in administrator ubuntu to be able to install php-cgi and do some CGI magic!${NC}"
-                fi; 
-                break;;
-        [Nn]* ) echo -e "Skipping php-cgi install - Ain't nobody got time for dat!"
-                break;;
-        * ) echo -e "${RED}Please press y or n.${NC}";;
-    esac
-done
+if ! [ -x "$(command -v php-cgi)" ];
+then
+    while true; do
+        read -p $'\033[01;35mDo you wish to install php-cgi (only available on Ubuntu w/ sudo rights)?: y/n \033[0m' yn
+        case $yn in
+            [Yy]* ) if [[ "$OSTYPE" == "linux-gnu"* ]] 
+                    then
+                        echo -e "Installing php-cgi..." &&
+                        sudo apt-get install php-cgi &&
+                        echo -e "php-cgi installed!"
+                    else
+                        echo -e "${RED}Please run this webserv in administrator ubuntu to be able to install php-cgi and do some CGI magic!${NC}"
+                    fi; 
+                    break;;
+            [Nn]* ) echo -e "Skipping php-cgi install - Ain't nobody got time for dat!"
+                    break;;
+            * ) echo -e "${RED}Please press y or n.${NC}";;
+        esac
+    done
+fi
+if [ ! -f ./scripts/host ]; then
+    echo -e "\n\n${MAGENTA}${BOLD}Please use your mighty root power to copy and paste the following lines to your /etc/hosts file:\n${NC}"
 
-echo -e "\n\n${MAGENTA}${BOLD}Please use your mighty root power to copy and paste the following lines to your /etc/hosts file:\n${NC}"
+    echo -e "####################################"
+    echo -e "##### <webserv sweatzer hosts> #####\n"
 
-echo -e "####################################"
-echo -e "##### <webserv sweatzer hosts> #####\n"
+    echo -e "127.0.0.1   sweatzer"
+    echo -e "127.0.0.1   sweatzer.com"
+    echo -e "127.0.0.1   www.sweatzer.com"
+    echo -e "127.0.0.1   sweatzersupasecret.com"
+    echo -e "127.0.0.1   www.sweatzersupasecret.com\n"
 
-echo -e "127.0.0.1   sweatzer"
-echo -e "127.0.0.1   sweatzer.com"
-echo -e "127.0.0.1   www.sweatzer.com"
-echo -e "127.0.0.1   sweatzersupasecret.com"
-echo -e "127.0.0.1   www.sweatzersupasecret.com\n"
+    echo -e "127.0.0.1   youpi"
+    echo -e "127.0.0.1   youpi.com"
+    echo -e "127.0.0.1   www.youpi.com\n"
 
-echo -e "127.0.0.1   youpi"
-echo -e "127.0.0.1   youpi.com"
-echo -e "127.0.0.1   www.youpi.com\n"
+    echo -e "127.0.0.1   acoudert.com"
+    echo -e "127.0.0.1   www.acoudert.com\n"
 
-echo -e "127.0.0.1   acoudert.com"
-echo -e "127.0.0.1   www.acoudert.com\n"
+    echo -e "##### </webserv sweatzer hosts> #####${NC}"
+    echo -e "#####################################\n"
 
-echo -e "##### </webserv sweatzer hosts> #####${NC}"
-echo -e "#####################################\n"
+    read -p $'\033[01;35m[...Please press enter once you\'re done...]\033[0m'
+    echo -e "I hope you rocked that copy pasting to the moon\n"
 
-read -p $'\033[01;35m[...Please press enter once you\'re done...]\033[0m'
-echo -e "I hope you rocked that copy pasting to the moon\n"
+    chmod 000 www/sweatzer/noperm
+    chmod 000 tests/acoudert_tester/pythonTest/noperm.html
 
-chmod 000 www/sweatzer/noperm
-chmod 000 tests/acoudert_tester/pythonTest/noperm.html
-
+    touch ./scripts/host
+    
+fi
